requirements.txt
alabaster==0.7.12
asgiref==3.2.7
Babel==2.8.0
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.8
django-allauth==0.41.0
django-analytical==2.6.0
django-ckeditor==5.9.0
django-crispy-forms==1.9.0
django-hitcount==1.3.0
django-js-asset==1.2.2
django-jsonfield==1.4.0
django-newsletter==0.9
django-ok-seo==0.8.7
django-robots==4.0
django-storages==1.9.1
django-taggit==1.2.0
dnspython==1.16.0
docutils==0.16
dropbox==10.1.1
idna==2.9
imagesize==1.2.0
Jinja2==2.11.2
ldif3==3.1.1
MarkupSafe==1.1.1
mock==4.0.2
oauthlib==3.1.0
packaging==20.4
Pillow==7.1.1
protobuf==3.6.1
psycopg2==2.8.5
Pygments==2.7.1
pyparsing==2.4.7
python-card-me==0.9.3
python-dateutil==2.6.0
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
snowballstemmer==2.0.0
sorl-thumbnail==12.6.3
Sphinx==3.2.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
surlex==0.2.0
unicodecsv==0.14.1
urllib3==1.25.9
